Rearrange decision code into own file
